added secret.h

+++ src/yesoulbike.cpp
@@ -167,7 +167,7 @@ void yesoulbike::characteristicChanged(const QLowEnergyCharacteristic &character
     if(ios_peloton_workaround && cadence && h && firstStateChanged)
     {
         h->virtualbike_setCadence(currentCrankRevolutions(),lastCrankEventTime());
-        h->virtualbike_setHeartRate(currentHeart().value());
+        h->virtualbike_setHeartRate((uint8_t)currentHeart().value());
     }
 #endif
 #endif
